Better parameter counting
The models all have more than one network. Count them separately and add up to
a total. Doesn't change the total for AudioLDM2 or Stable Audio, but does
reveal that MMAudio is about twice as large as what we previously counted
(because only the main DiT was being counted).

Note that this is still a bit error prone, because how to walk the data
structures and find all the networks is different for each model.